LayoutTests/platform/ios-sim-deprecated/accessibility/progressbar.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
<script>
var successfullyParsed = false;
</script>
</head>
<body id="body">

<div id="progressbar" role="progressbar" aria-label="Popularity" aria-valuemin="1" aria-valuenow="5" aria-valuemax="12">blah</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that progress bars show up in the AX hierarchy and that min, current and max values are reported correctly.");

    if (window.accessibilityController) {

        var button1 = accessibilityController.accessibleElementById("progressbar");
        shouldBe("button1.iphoneLabel", "'AXLabel: Popularity'");
        shouldBe("button1.iphoneValue", "'AXValue: 5.00'");
        shouldBeTrue("button1.iphoneIsElement");
        shouldBe("button1.minValue", "1");
        shouldBe("button1.maxValue", "12");
    }

    successfullyParsed = true;
</script>

<script src="../../../resources/js-test-post.js"></script>

</body>
</html>
